+++ client/app/src/components/CheckList/Write/Section.tsx
@@ -5,34 +5,6 @@ import * as Styled from "./style.ts"
 import H1 from "@/components/Common/Font/Heading/H1";
 import ChecklistItem from "@/components/CheckList/Write/CheckListItem.tsx";
 
-/*
-interface SectionProps {
-    section: TSection;
-    onItemCheck: (sectionId: string, subSectionName: string, itemId: string) => void;
-}
-
-const Section: React.FC<SectionProps> = ({ section, onItemCheck }) => {
-    return (
-        <div>
-            <Styled.StyledSection>
-                <H1 text={section.name} />
-            </Styled.StyledSection>
-            {Object.keys(section.subSections).map(subSectionName => (
-                <React.Fragment key={subSectionName}>
-                    <SubSection
-                        sectionId={section.id}
-                        subSectionName={subSectionName}
-                        items={section.subSections[subSectionName]}
-                        onItemCheck={onItemCheck}
-                    />
-                    <hr/>
-                </React.Fragment>
-            ))}
-        </div>
-    );
-};
-
- */
 interface SectionProps {
     section: TSection;
     onItemCheck: (sectionName: string, subSectionName: string | null, detailSectionName: string | null, checked: boolean) => void;
